+++ index.html
@@ -134,7 +134,7 @@
     </li>
   </template>
 
-  <script src="script.js"></script>
+  <script type="module" src="./src/components/index.js"></script>
 </body>
 
 </html>
